Remove/rename some unused Chromely example code
