+++ src/Programs/ChromelyBlazor.Shared/GlobalUsings.cs
@@ -1,15 +1,12 @@
 global using System.IO.MemoryMappedFiles;
 global using System.Net;
 global using System.Net.NetworkInformation;
-global using System.Text.Json;
 
 global using Chromely.Core.Configuration;
 global using Chromely.Core.Infrastructure;
 global using Chromely.Core.Host;
 global using Chromely.Core.Logging;
 global using Chromely.Core.Network;
-global using ChromelyBlazor.Shared.Models;
 
 global using Microsoft.Extensions.Logging;
-global using Microsoft.JSInterop;
